:ambulance: fix(child_clinical_measurements_form_validator): Improve child's measurement validations
Refactored the code in child_clinical_measurements_form_validator to include a completion check for skin fold measurements for each visit. This is important to distinguish unmeasured data from not applicable cases. Also, the verbose names of measurement fields have been updated for better readability and understanding, with changes reflected in the form validator and admin files for proper validation and display.