Remove skin fold validation from child clinical measurements
In the form validator class for child clinical measurements, previously there was a check in place for the skin fold measurements. However, due to a change in the research protocol, the skin fold measurements are no longer required. Those specific validations are thus removed to be in line with the new protocol.