(testing): Add validation test for Infant ARV Exposure Form
This commit adds new test cases to enhance the test coverage of the Infant ARV Exposure Form's validation function. Specifically, it checks the correct validation for 'azt_within_72h' and 'snvp_dose_within_72h' fields as well as handles the case when 'nvp_cont_dosing' field is YES but 'sdnvp_after_birth' is NO. Also, the order of class and constant imports has been modified in the test script for better readability.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>